Moved and reorganized common library to separate templates from typescript code
