+++ projects/common/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@lowcodeunit/lcu-charts-common",
-	"version": "1.1.0",
+	"version": "1.2.0",
   "description": "A Charting and Graphing Library based on and influenced by Ngx-charts",
   "repository": {
     "type": "git",
