+++ projects/common/src/lib/controls/area-chart/area-stacked/area-chart-stacked.component.html
@@ -1,0 +1,129 @@
+<lcu-charts-chart
+  [view]="[width, height]"
+  [showLegend]="legend"
+  [legendOptions]="legendOptions"
+  [activeEntries]="activeEntries"
+  [animations]="animations"
+  (legendLabelClick)="onClick($event)"
+  (legendLabelActivate)="onActivate($event)"
+  (legendLabelDeactivate)="onDeactivate($event)">
+  <svg:defs>
+    <svg:clipPath [attr.id]="clipPathId">
+      <svg:rect
+        [attr.width]="dims.width + 10"
+        [attr.height]="dims.height + 10"
+        [attr.transform]="'translate(-5, -5)'"
+      />
+    </svg:clipPath>
+  </svg:defs>
+  <svg:g [attr.transform]="transform" class="area-chart chart">
+    <svg:g
+      lcu-charts-x-axis
+      *ngIf="xAxis"
+      [xScale]="xScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showXAxisLabel"
+      [labelText]="xAxisLabel"
+      [trimTicks]="trimXAxisTicks"
+      [rotateTicks]="rotateXAxisTicks"
+      [maxTickLength]="maxXAxisTickLength"
+      [tickFormatting]="xAxisTickFormatting"
+      [ticks]="xAxisTicks"
+      (dimensionsChanged)="updateXAxisHeight($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-y-axis
+      *ngIf="yAxis"
+      [yScale]="yScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showYAxisLabel"
+      [labelText]="yAxisLabel"
+      [trimTicks]="trimYAxisTicks"
+      [maxTickLength]="maxYAxisTickLength"
+      [tickFormatting]="yAxisTickFormatting"
+      [ticks]="yAxisTicks"
+      (dimensionsChanged)="updateYAxisWidth($event)"
+    ></svg:g>
+    <svg:g [attr.clip-path]="clipPath">
+      <svg:g *ngFor="let series of results; trackBy: trackBy">
+        <svg:g
+          lcu-charts-area-series
+          [xScale]="xScale"
+          [yScale]="yScale"
+          [colors]="colors"
+          [data]="series"
+          [scaleType]="scaleType"
+          [gradient]="gradient"
+          [activeEntries]="activeEntries"
+          stacked="true"
+          [curve]="curve"
+          [animations]="animations"
+        />
+      </svg:g>
+
+      <svg:g *ngIf="!tooltipDisabled" (mouseleave)="hideCircles()">
+        <svg:g
+          lcu-charts-tooltip-area
+          [dims]="dims"
+          [xSet]="xSet"
+          [xScale]="xScale"
+          [yScale]="yScale"
+          [results]="results"
+          [colors]="colors"
+          [tooltipDisabled]="tooltipDisabled"
+          [tooltipTemplate]="seriesTooltipTemplate"
+          (hover)="updateHoveredVertical($event)"
+        />
+
+        <svg:g *ngFor="let series of results; trackBy: trackBy">
+          <svg:g
+            lcu-charts-circle-series
+            type="stacked"
+            [xScale]="xScale"
+            [yScale]="yScale"
+            [colors]="colors"
+            [activeEntries]="activeEntries"
+            [data]="series"
+            [scaleType]="scaleType"
+            [visibleValue]="hoveredVertical"
+            [tooltipDisabled]="tooltipDisabled"
+            [tooltipTemplate]="tooltipTemplate"
+            (select)="onClick($event, series)"
+            (activate)="onActivate($event)"
+            (deactivate)="onDeactivate($event)"
+          />
+        </svg:g>
+      </svg:g>
+    </svg:g>
+  </svg:g>
+  <svg:g
+    lcu-charts-timeline
+    *ngIf="timeline && scaleType != 'ordinal'"
+    [attr.transform]="timelineTransform"
+    [results]="results"
+    [view]="[timelineWidth, height]"
+    [height]="timelineHeight"
+    [scheme]="scheme"
+    [customColors]="customColors"
+    [legend]="legend"
+    [scaleType]="scaleType"
+    (onDomainChange)="updateDomain($event)"
+  >
+    <svg:g *ngFor="let series of results; trackBy: trackBy">
+      <svg:g
+        lcu-charts-area-series
+        [xScale]="timelineXScale"
+        [yScale]="timelineYScale"
+        [colors]="colors"
+        [data]="series"
+        [scaleType]="scaleType"
+        [gradient]="gradient"
+        stacked="true"
+        [curve]="curve"
+        [animations]="animations"
+      />
+    </svg:g>
+  </svg:g>
+</lcu-charts-chart>
