+++ projects/common/src/lib/controls/bar-chart/bar-chart.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { ChartCommonModule } from '../../common/chart-common.module';
 import { BarComponent } from './bar.component';
-import { BarHorizontalComponent } from './bar-horizontal.component';
-import { BarHorizontal2DComponent } from './bar-horizontal-2d.component';
-import { BarHorizontalNormalizedComponent } from './bar-horizontal-normalized.component';
-import { BarHorizontalStackedComponent } from './bar-horizontal-stacked.component';
-import { BarVerticalComponent } from './bar-vertical.component';
-import { BarVertical2DComponent } from './bar-vertical-2d.component';
-import { BarVerticalNormalizedComponent } from './bar-vertical-normalized.component';
-import { BarVerticalStackedComponent } from './bar-vertical-stacked.component';
+import { BarHorizontalSimpleComponent } from './bar-horizontal-simple/bar-horizontal-simple.component';
+import { BarHorizontalGroupedComponent } from './bar-horizontal-grouped/bar-horizontal-grouped.component';
+import { BarHorizontalNormalizedComponent } from './bar-horizontal-normalized/bar-horizontal-normalized.component';
+import { BarHorizontalStackedComponent } from './bar-horizontal-stacked/bar-horizontal-stacked.component';
+import { BarVerticalSimpleComponent } from './bar-vertical-simple/bar-vertical-simple.component';
+import { BarVerticalGroupedComponent } from './bar-vertical-grouped/bar-vertical-grouped.component';
+import { BarVerticalNormalizedComponent } from './bar-vertical-normalized/bar-vertical-normalized.component';
+import { BarVerticalStackedComponent } from './bar-vertical-stacked/bar-vertical-stacked.component';
 import { SeriesHorizontal } from './series-horizontal.component';
 import { SeriesVerticalComponent } from './series-vertical.component';
 import { BarLabelComponent } from './bar-label.component';
@@ -19,12 +19,12 @@ import { BarLabelComponent } from './bar-label.component';
   ],
   declarations: [
     BarComponent,
-    BarHorizontalComponent,
-    BarHorizontal2DComponent,
+    BarHorizontalSimpleComponent,
+    BarHorizontalGroupedComponent,
     BarHorizontalNormalizedComponent,
     BarHorizontalStackedComponent,
-    BarVerticalComponent,
-    BarVertical2DComponent,
+    BarVerticalSimpleComponent,
+    BarVerticalGroupedComponent,
     BarVerticalNormalizedComponent,
     BarVerticalStackedComponent,
     BarLabelComponent,
@@ -33,12 +33,12 @@ import { BarLabelComponent } from './bar-label.component';
   ],
   exports: [
     BarComponent,
-    BarHorizontalComponent,
-    BarHorizontal2DComponent,
+    BarHorizontalSimpleComponent,
+    BarHorizontalGroupedComponent,
     BarHorizontalNormalizedComponent,
     BarHorizontalStackedComponent,
-    BarVerticalComponent,
-    BarVertical2DComponent,
+    BarVerticalSimpleComponent,
+    BarVerticalGroupedComponent,
     BarVerticalNormalizedComponent,
     BarVerticalStackedComponent,
     BarLabelComponent,
