+++ projects/common/src/lib/controls/bar-chart/bar-horizontal-normalized/bar-horizontal-normalized.component.html
@@ -1,0 +1,66 @@
+<lcu-charts-chart
+  [view]="[width, height]"
+  [showLegend]="legend"
+  [legendOptions]="legendOptions"
+  [activeEntries]="activeEntries"
+  [animations]="animations"
+  (legendLabelActivate)="onActivate($event, undefined, true)"
+  (legendLabelDeactivate)="onDeactivate($event, undefined, true)"
+  (legendLabelClick)="onClick($event)"
+  >
+  <svg:g [attr.transform]="transform" class="bar-chart chart">
+    <svg:g
+      lcu-charts-x-axis
+      *ngIf="xAxis"
+      [xScale]="xScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showXAxisLabel"
+      [labelText]="xAxisLabel"
+      [trimTicks]="trimXAxisTicks"
+      [rotateTicks]="rotateXAxisTicks"
+      [maxTickLength]="maxXAxisTickLength"
+      [tickFormatting]="xAxisTickFormatting"
+      [ticks]="xAxisTicks"
+      (dimensionsChanged)="updateXAxisHeight($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-y-axis
+      *ngIf="yAxis"
+      [yScale]="yScale"
+      [dims]="dims"
+      [showLabel]="showYAxisLabel"
+      [labelText]="yAxisLabel"
+      [trimTicks]="trimYAxisTicks"
+      [maxTickLength]="maxYAxisTickLength"
+      [tickFormatting]="yAxisTickFormatting"
+      [ticks]="yAxisTicks"
+      (dimensionsChanged)="updateYAxisWidth($event)"
+    ></svg:g>
+    <svg:g
+      *ngFor="let group of results; trackBy: trackBy"
+      [@animationState]="'active'"
+      [attr.transform]="groupTransform(group)"
+    >
+      <svg:g
+        lcu-charts-series-horizontal
+        type="normalized"
+        [xScale]="xScale"
+        [yScale]="yScale"
+        [activeEntries]="activeEntries"
+        [colors]="colors"
+        [series]="group.series"
+        [dims]="dims"
+        [gradient]="gradient"
+        [tooltipDisabled]="tooltipDisabled"
+        [tooltipTemplate]="tooltipTemplate"
+        [seriesName]="group.name"
+        [animations]="animations"
+        (select)="onClick($event, group)"
+        (activate)="onActivate($event, group)"
+        (deactivate)="onDeactivate($event, group)"
+        [noBarWhenZero]="noBarWhenZero"
+      />
+    </svg:g>
+  </svg:g>
+</lcu-charts-chart>
