+++ projects/common/src/lib/controls/bar-chart/bar-horizontal-simple/bar-horizontal-simple.component.html
@@ -1,0 +1,62 @@
+<lcu-charts-chart
+  [view]="[width, height]"
+  [showLegend]="legend"
+  [legendOptions]="legendOptions"
+  [activeEntries]="activeEntries"
+  [animations]="animations"
+  (legendLabelClick)="onClick($event)"
+  (legendLabelActivate)="onActivate($event, true)"
+  (legendLabelDeactivate)="onDeactivate($event, true)">
+  <svg:g [attr.transform]="transform" class="bar-chart chart">
+    <svg:g
+      lcu-charts-x-axis
+      *ngIf="xAxis"
+      [xScale]="xScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showXAxisLabel"
+      [labelText]="xAxisLabel"
+      [trimTicks]="trimXAxisTicks"
+      [rotateTicks]="rotateXAxisTicks"
+      [maxTickLength]="maxXAxisTickLength"
+      [tickFormatting]="xAxisTickFormatting"
+      [ticks]="xAxisTicks"
+      (dimensionsChanged)="updateXAxisHeight($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-y-axis
+      *ngIf="yAxis"
+      [yScale]="yScale"
+      [dims]="dims"
+      [showLabel]="showYAxisLabel"
+      [labelText]="yAxisLabel"
+      [trimTicks]="trimYAxisTicks"
+      [maxTickLength]="maxYAxisTickLength"
+      [tickFormatting]="yAxisTickFormatting"
+      [ticks]="yAxisTicks"
+      [yAxisOffset]="dataLabelMaxWidth.negative"
+      (dimensionsChanged)="updateYAxisWidth($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-series-horizontal
+      [xScale]="xScale"
+      [yScale]="yScale"
+      [colors]="colors"
+      [series]="results"
+      [dims]="dims"
+      [gradient]="gradient"
+      [tooltipDisabled]="tooltipDisabled"
+      [tooltipTemplate]="tooltipTemplate"
+      [activeEntries]="activeEntries"
+      [roundEdges]="roundEdges"
+      [animations]="animations"
+      [showDataLabel]="showDataLabel"
+      [dataLabelFormatting]="dataLabelFormatting"
+      [noBarWhenZero]="noBarWhenZero"
+      (select)="onClick($event)"
+      (activate)="onActivate($event)"
+      (deactivate)="onDeactivate($event)"
+      (dataLabelWidthChanged)="onDataLabelMaxWidthChanged($event)"
+    ></svg:g>
+  </svg:g>
+</lcu-charts-chart>
