+++ projects/common/src/lib/controls/bar-chart/bar-horizontal-simple/bar-horizontal-simple.component.ts
@@ -1,92 +1,26 @@
 import {
   Component,
   Input,
-  ViewEncapsulation,
   Output,
   EventEmitter,
+  ViewEncapsulation,
   ChangeDetectionStrategy,
   ContentChild,
   TemplateRef
 } from '@angular/core';
 import { scaleBand, scaleLinear } from 'd3-scale';
-
-import { calculateViewDimensions, ViewDimensions } from '../../common/view-dimensions.helper';
-import { ColorHelper } from '../../common/color.helper';
-import { BaseChartComponent } from '../../common/base-chart.component';
-import { DataItem } from '../../models/chart-data.model';
+import { calculateViewDimensions, ViewDimensions } from '../../../common/view-dimensions.helper';
+import { ColorHelper } from '../../../common/color.helper';
+import { BaseChartComponent } from '../../../common/base-chart.component';
 
 @Component({
-  selector: 'lcu-charts-bar-vertical',
-  template: `
-    <lcu-charts-chart
-      [view]="[width, height]"
-      [showLegend]="legend"
-      [legendOptions]="legendOptions"
-      [activeEntries]="activeEntries"
-      [animations]="animations"
-      (legendLabelClick)="onClick($event)"
-      (legendLabelActivate)="onActivate($event, true)"
-      (legendLabelDeactivate)="onDeactivate($event, true)"
-    >
-      <svg:g [attr.transform]="transform" class="bar-chart chart">
-        <svg:g
-          lcu-charts-x-axis
-          *ngIf="xAxis"
-          [xScale]="xScale"
-          [dims]="dims"
-          [showLabel]="showXAxisLabel"
-          [labelText]="xAxisLabel"
-          [trimTicks]="trimXAxisTicks"
-          [rotateTicks]="rotateXAxisTicks"
-          [maxTickLength]="maxXAxisTickLength"
-          [tickFormatting]="xAxisTickFormatting"
-          [ticks]="xAxisTicks"
-          [xAxisOffset]="dataLabelMaxHeight.negative"
-          (dimensionsChanged)="updateXAxisHeight($event)"
-        ></svg:g>
-        <svg:g
-          lcu-charts-y-axis
-          *ngIf="yAxis"
-          [yScale]="yScale"
-          [dims]="dims"
-          [showGridLines]="showGridLines"
-          [showLabel]="showYAxisLabel"
-          [labelText]="yAxisLabel"
-          [trimTicks]="trimYAxisTicks"
-          [maxTickLength]="maxYAxisTickLength"
-          [tickFormatting]="yAxisTickFormatting"
-          [ticks]="yAxisTicks"
-          (dimensionsChanged)="updateYAxisWidth($event)"
-        ></svg:g>
-        <svg:g
-          lcu-charts-series-vertical
-          [xScale]="xScale"
-          [yScale]="yScale"
-          [colors]="colors"
-          [series]="results"
-          [dims]="dims"
-          [gradient]="gradient"
-          [tooltipDisabled]="tooltipDisabled"
-          [tooltipTemplate]="tooltipTemplate"
-          [showDataLabel]="showDataLabel"
-          [dataLabelFormatting]="dataLabelFormatting"
-          [activeEntries]="activeEntries"
-          [roundEdges]="roundEdges"
-          [animations]="animations"
-          [noBarWhenZero]="noBarWhenZero"
-          (activate)="onActivate($event)"
-          (deactivate)="onDeactivate($event)"
-          (select)="onClick($event)"
-          (dataLabelHeightChanged)="onDataLabelMaxHeightChanged($event)"
-        ></svg:g>
-      </svg:g>
-    </lcu-charts-chart>
-  `,
+  selector: 'lcu-charts-bar-horizontal-simple',
+  templateUrl: 'bar-horizontal-simple.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  styleUrls: ['../../common/base-chart.component.scss'],
+  styleUrls: ['../../../common/base-chart.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class BarVerticalComponent extends BaseChartComponent {
+export class BarHorizontalSimpleComponent extends BaseChartComponent {
   @Input() legend = false;
   @Input() legendTitle: string = 'Legend';
   @Input() legendPosition: string = 'right';
@@ -113,8 +47,8 @@ export class BarVerticalComponent extends BaseChartComponent {
   @Input() barPadding = 8;
   @Input() roundDomains: boolean = false;
   @Input() roundEdges: boolean = true;
-  @Input() yScaleMax: number;
-  @Input() yScaleMin: number;
+  @Input() xScaleMax: number;
+  @Input() xScaleMin: number;
   @Input() showDataLabel: boolean = false;
   @Input() dataLabelFormatting: any;
   @Input() noBarWhenZero: boolean = true;
@@ -125,25 +59,26 @@ export class BarVerticalComponent extends BaseChartComponent {
   @ContentChild('tooltipTemplate', { static: false }) tooltipTemplate: TemplateRef<any>;
 
   dims: ViewDimensions;
-  xScale: any;
   yScale: any;
+  xScale: any;
   xDomain: any;
   yDomain: any;
   transform: string;
   colors: ColorHelper;
-  margin: any[] = [10, 20, 10, 20];
+  margin = [10, 20, 10, 20];
   xAxisHeight: number = 0;
   yAxisWidth: number = 0;
   legendOptions: any;
-  dataLabelMaxHeight: any = { negative: 0, positive: 0 };
+  dataLabelMaxWidth: any = { negative: 0, positive: 0 };
 
   update(): void {
     super.update();
 
     if (!this.showDataLabel) {
-      this.dataLabelMaxHeight = { negative: 0, positive: 0 };
+      this.dataLabelMaxWidth = { negative: 0, positive: 0 };
     }
-    this.margin = [10 + this.dataLabelMaxHeight.positive, 20, 10 + this.dataLabelMaxHeight.negative, 20];
+
+    this.margin = [10, 20 + this.dataLabelMaxWidth.positive, 10, 20 + this.dataLabelMaxWidth.negative];
 
     this.dims = calculateViewDimensions({
       width: this.width,
@@ -162,64 +97,57 @@ export class BarVerticalComponent extends BaseChartComponent {
 
     this.formatDates();
 
-    if (this.showDataLabel) {
-      this.dims.height -= this.dataLabelMaxHeight.negative;
-    }
     this.xScale = this.getXScale();
     this.yScale = this.getYScale();
 
     this.setColors();
     this.legendOptions = this.getLegendOptions();
 
-    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0] + this.dataLabelMaxHeight.negative})`;
+    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0]})`;
   }
 
   getXScale(): any {
     this.xDomain = this.getXDomain();
-    const spacing = this.xDomain.length / (this.dims.width / this.barPadding + 1);
-    return scaleBand()
+
+    const scale = scaleLinear()
       .range([0, this.dims.width])
-      .paddingInner(spacing)
       .domain(this.xDomain);
+
+    return this.roundDomains ? scale.nice() : scale;
   }
 
   getYScale(): any {
     this.yDomain = this.getYDomain();
-    const scale = scaleLinear()
-      .range([this.dims.height, 0])
+    const spacing = this.yDomain.length / (this.dims.height / this.barPadding + 1);
+
+    return scaleBand()
+      .rangeRound([0, this.dims.height])
+      .paddingInner(spacing)
       .domain(this.yDomain);
-    return this.roundDomains ? scale.nice() : scale;
   }
 
   getXDomain(): any[] {
-    return this.results.map(d => d.label);
-  }
-
-  getYDomain(): [number, number] {
     const values = this.results.map(d => d.value);
+    const min = this.xScaleMin ? Math.min(this.xScaleMin, ...values) : Math.min(0, ...values);
 
-    let min = this.yScaleMin ? Math.min(this.yScaleMin, ...values) : Math.min(0, ...values);
-    if (this.yAxisTicks && !this.yAxisTicks.some(isNaN)) {
-      min = Math.min(min, ...this.yAxisTicks);
-    }
-
-    let max = this.yScaleMax ? Math.max(this.yScaleMax, ...values) : Math.max(0, ...values);
-    if (this.yAxisTicks && !this.yAxisTicks.some(isNaN)) {
-      max = Math.max(max, ...this.yAxisTicks);
-    }
+    const max = this.xScaleMax ? Math.max(this.xScaleMax, ...values) : Math.max(0, ...values);
     return [min, max];
   }
 
-  onClick(data: DataItem) {
+  getYDomain(): any[] {
+    return this.results.map(d => d.label);
+  }
+
+  onClick(data): void {
     this.select.emit(data);
   }
 
   setColors(): void {
     let domain;
     if (this.schemeType === 'ordinal') {
-      domain = this.xDomain;
-    } else {
       domain = this.yDomain;
+    } else {
+      domain = this.xDomain;
     }
 
     this.colors = new ColorHelper(this.scheme, this.schemeType, domain, this.customColors);
@@ -234,13 +162,14 @@ export class BarVerticalComponent extends BaseChartComponent {
       position: this.legendPosition
     };
     if (opts.scaleType === 'ordinal') {
-      opts.domain = this.xDomain;
+      opts.domain = this.yDomain;
       opts.colors = this.colors;
       opts.title = this.legendTitle;
     } else {
-      opts.domain = this.yDomain;
+      opts.domain = this.xDomain;
       opts.colors = this.colors.scale;
     }
+
     return opts;
   }
 
@@ -254,11 +183,11 @@ export class BarVerticalComponent extends BaseChartComponent {
     this.update();
   }
 
-  onDataLabelMaxHeightChanged(event) {
+  onDataLabelMaxWidthChanged(event) {
     if (event.size.negative) {
-      this.dataLabelMaxHeight.negative = Math.max(this.dataLabelMaxHeight.negative, event.size.height);
+      this.dataLabelMaxWidth.negative = Math.max(this.dataLabelMaxWidth.negative, event.size.width);
     } else {
-      this.dataLabelMaxHeight.positive = Math.max(this.dataLabelMaxHeight.positive, event.size.height);
+      this.dataLabelMaxWidth.positive = Math.max(this.dataLabelMaxWidth.positive, event.size.width);
     }
     if (event.index === this.results.length - 1) {
       setTimeout(() => this.update());
