+++ projects/common/src/lib/controls/bar-chart/bar-vertical-grouped/bar-vertical-grouped.component.html
@@ -1,0 +1,74 @@
+<lcu-charts-chart
+  [view]="[width, height]"
+  [showLegend]="legend"
+  [legendOptions]="legendOptions"
+  [activeEntries]="activeEntries"
+  [animations]="animations"
+  (legendLabelActivate)="onActivate($event, undefined, true)"
+  (legendLabelDeactivate)="onDeactivate($event, undefined, true)"
+  (legendLabelClick)="onClick($event)">
+  <svg:g [attr.transform]="transform" class="bar-chart chart">
+    <svg:g
+      lcu-charts-grid-panel-series
+      [xScale]="groupScale"
+      [yScale]="valueScale"
+      [data]="results"
+      [dims]="dims"
+      orient="vertical"
+    ></svg:g>
+    <svg:g
+      lcu-charts-x-axis
+      *ngIf="xAxis"
+      [xScale]="groupScale"
+      [dims]="dims"
+      [showLabel]="showXAxisLabel"
+      [labelText]="xAxisLabel"
+      [trimTicks]="trimXAxisTicks"
+      [rotateTicks]="rotateXAxisTicks"
+      [maxTickLength]="maxXAxisTickLength"
+      [tickFormatting]="xAxisTickFormatting"
+      [ticks]="xAxisTicks"
+      [xAxisOffset]="dataLabelMaxHeight.negative"
+      (dimensionsChanged)="updateXAxisHeight($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-y-axis
+      *ngIf="yAxis"
+      [yScale]="valueScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showYAxisLabel"
+      [labelText]="yAxisLabel"
+      [trimTicks]="trimYAxisTicks"
+      [maxTickLength]="maxYAxisTickLength"
+      [tickFormatting]="yAxisTickFormatting"
+      [ticks]="yAxisTicks"
+      (dimensionsChanged)="updateYAxisWidth($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-series-vertical
+      *ngFor="let group of results; let index = index; trackBy: trackBy"
+      [@animationState]="'active'"
+      [attr.transform]="groupTransform(group)"
+      [activeEntries]="activeEntries"
+      [xScale]="innerScale"
+      [yScale]="valueScale"
+      [colors]="colors"
+      [series]="group.series"
+      [dims]="dims"
+      [gradient]="gradient"
+      [tooltipDisabled]="tooltipDisabled"
+      [tooltipTemplate]="tooltipTemplate"
+      [showDataLabel]="showDataLabel"
+      [dataLabelFormatting]="dataLabelFormatting"
+      [seriesName]="group.name"
+      [roundEdges]="roundEdges"
+      [animations]="animations"
+      [noBarWhenZero]="noBarWhenZero"
+      (select)="onClick($event, group)"
+      (activate)="onActivate($event, group)"
+      (deactivate)="onDeactivate($event, group)"
+      (dataLabelHeightChanged)="onDataLabelMaxHeightChanged($event, index)"
+    />
+  </svg:g>
+</lcu-charts-chart>
