+++ projects/common/src/lib/controls/bar-chart/bar-vertical-normalized/bar-vertical-normalized.component.ts
@@ -11,85 +11,14 @@ import {
 import { trigger, style, animate, transition } from '@angular/animations';
 import { scaleBand, scaleLinear } from 'd3-scale';
 
-import { calculateViewDimensions, ViewDimensions } from '../../common/view-dimensions.helper';
-import { ColorHelper } from '../../common/color.helper';
-import { BaseChartComponent } from '../../common/base-chart.component';
+import { calculateViewDimensions, ViewDimensions } from '../../../common/view-dimensions.helper';
+import { ColorHelper } from '../../../common/color.helper';
+import { BaseChartComponent } from '../../../common/base-chart.component';
 
 @Component({
-  selector: 'lcu-charts-bar-vertical-stacked',
-  template: `
-    <lcu-charts-chart
-      [view]="[width, height]"
-      [showLegend]="legend"
-      [legendOptions]="legendOptions"
-      [activeEntries]="activeEntries"
-      [animations]="animations"
-      (legendLabelActivate)="onActivate($event, undefined, true)"
-      (legendLabelDeactivate)="onDeactivate($event, undefined, true)"
-      (legendLabelClick)="onClick($event)"
-    >
-      <svg:g [attr.transform]="transform" class="bar-chart chart">
-        <svg:g
-          lcu-charts-x-axis
-          *ngIf="xAxis"
-          [xScale]="xScale"
-          [dims]="dims"
-          [showLabel]="showXAxisLabel"
-          [labelText]="xAxisLabel"
-          [trimTicks]="trimXAxisTicks"
-          [rotateTicks]="rotateXAxisTicks"
-          [maxTickLength]="maxXAxisTickLength"
-          [tickFormatting]="xAxisTickFormatting"
-          [ticks]="xAxisTicks"
-          [xAxisOffset]="dataLabelMaxHeight.negative"
-          (dimensionsChanged)="updateXAxisHeight($event)"
-        ></svg:g>
-        <svg:g
-          lcu-charts-y-axis
-          *ngIf="yAxis"
-          [yScale]="yScale"
-          [dims]="dims"
-          [showGridLines]="showGridLines"
-          [showLabel]="showYAxisLabel"
-          [labelText]="yAxisLabel"
-          [trimTicks]="trimYAxisTicks"
-          [maxTickLength]="maxYAxisTickLength"
-          [tickFormatting]="yAxisTickFormatting"
-          [ticks]="yAxisTicks"
-          (dimensionsChanged)="updateYAxisWidth($event)"
-        ></svg:g>
-        <svg:g
-          *ngFor="let group of results; let index = index; trackBy: trackBy"
-          [@animationState]="'active'"
-          [attr.transform]="groupTransform(group)"
-        >
-          <svg:g
-            lcu-charts-series-vertical
-            type="stacked"
-            [xScale]="xScale"
-            [yScale]="yScale"
-            [activeEntries]="activeEntries"
-            [colors]="colors"
-            [series]="group.series"
-            [dims]="dims"
-            [gradient]="gradient"
-            [tooltipDisabled]="tooltipDisabled"
-            [tooltipTemplate]="tooltipTemplate"
-            [showDataLabel]="showDataLabel"
-            [dataLabelFormatting]="dataLabelFormatting"
-            [seriesName]="group.name"
-            [animations]="animations"
-            [noBarWhenZero]="noBarWhenZero"
-            (select)="onClick($event, group)"
-            (activate)="onActivate($event, group)"
-            (deactivate)="onDeactivate($event, group)"
-            (dataLabelHeightChanged)="onDataLabelMaxHeightChanged($event, index)"
-          />
-        </svg:g>
-      </svg:g>
-    </lcu-charts-chart>
-  `,
-  styleUrls: ['../../common/base-chart.component.scss'],
+  selector: 'lcu-charts-bar-vertical-normalized',
+  templateUrl: './bar-vertical-normalized.component.html',
+  styleUrls: ['../../../common/base-chart.component.scss'],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
   animations: [
@@ -104,7 +33,7 @@ import { BaseChartComponent } from '../../common/base-chart.component';
     ])
   ]
 })
-export class BarVerticalStackedComponent extends BaseChartComponent {
+export class BarVerticalNormalizedComponent extends BaseChartComponent {
   @Input() legend = false;
   @Input() legendTitle: string = 'Legend';
   @Input() legendPosition: string = 'right';
@@ -130,9 +59,6 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
   @Input() yAxisTicks: any[];
   @Input() barPadding = 8;
   @Input() roundDomains: boolean = false;
-  @Input() yScaleMax: number;
-  @Input() showDataLabel: boolean = false;
-  @Input() dataLabelFormatting: any;
   @Input() noBarWhenZero: boolean = true;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
@@ -147,21 +73,14 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
   xScale: any;
   yScale: any;
   transform: string;
-  tickFormatting: (label: string) => string;
   colors: ColorHelper;
   margin = [10, 20, 10, 20];
   xAxisHeight: number = 0;
   yAxisWidth: number = 0;
   legendOptions: any;
-  dataLabelMaxHeight: any = { negative: 0, positive: 0 };
 
   update(): void {
     super.update();
-
-    if (!this.showDataLabel) {
-      this.dataLabelMaxHeight = { negative: 0, positive: 0 };
-    }
-    this.margin = [10 + this.dataLabelMaxHeight.positive, 20, 10 + this.dataLabelMaxHeight.negative, 20];
 
     this.dims = calculateViewDimensions({
       width: this.width,
@@ -178,10 +97,6 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
       legendPosition: this.legendPosition
     });
 
-    if (this.showDataLabel) {
-      this.dims.height -= this.dataLabelMaxHeight.negative;
-    }
-
     this.formatDates();
 
     this.groupDomain = this.getGroupDomain();
@@ -194,7 +109,7 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
     this.setColors();
     this.legendOptions = this.getLegendOptions();
 
-    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0] + this.dataLabelMaxHeight.negative})`;
+    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0]})`;
   }
 
   getGroupDomain() {
@@ -204,6 +119,7 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
         domain.push(group.label);
       }
     }
+
     return domain;
   }
 
@@ -216,38 +132,17 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
         }
       }
     }
+
     return domain;
   }
 
   getValueDomain() {
-    const domain = [];
-    let smallest = 0;
-    let biggest = 0;
-    for (const group of this.results) {
-      let smallestSum = 0;
-      let biggestSum = 0;
-      for (const d of group.series) {
-        if (d.value < 0) {
-          smallestSum += d.value;
-        } else {
-          biggestSum += d.value;
-        }
-        smallest = d.value < smallest ? d.value : smallest;
-        biggest = d.value > biggest ? d.value : biggest;
-      }
-      domain.push(smallestSum);
-      domain.push(biggestSum);
-    }
-    domain.push(smallest);
-    domain.push(biggest);
-
-    const min = Math.min(0, ...domain);
-    const max = this.yScaleMax ? Math.max(this.yScaleMax, ...domain) : Math.max(...domain);
-    return [min, max];
+    return [0, 100];
   }
 
   getXScale(): any {
     const spacing = this.groupDomain.length / (this.dims.width / this.barPadding + 1);
+
     return scaleBand()
       .rangeRound([0, this.dims.width])
       .paddingInner(spacing)
@@ -261,19 +156,8 @@ export class BarVerticalStackedComponent extends BaseChartComponent {
     return this.roundDomains ? scale.nice() : scale;
   }
 
-  onDataLabelMaxHeightChanged(event, groupIndex) {
-    if (event.size.negative) {
-      this.dataLabelMaxHeight.negative = Math.max(this.dataLabelMaxHeight.negative, event.size.height);
-    } else {
-      this.dataLabelMaxHeight.positive = Math.max(this.dataLabelMaxHeight.positive, event.size.height);
-    }
-    if (groupIndex === this.results.length - 1) {
-      setTimeout(() => this.update());
-    }
-  }
-
   groupTransform(group) {
-    return `translate(${this.xScale(group.name) || 0}, 0)`;
+    return `translate(${this.xScale(group.name)}, 0)`;
   }
 
   onClick(data, group?) {
