+++ projects/common/src/lib/controls/bar-chart/bar-vertical-stacked/bar-vertical-stacked.component.html
@@ -1,0 +1,69 @@
+<lcu-charts-chart
+  [view]="[width, height]"
+  [showLegend]="legend"
+  [legendOptions]="legendOptions"
+  [activeEntries]="activeEntries"
+  [animations]="animations"
+  (legendLabelActivate)="onActivate($event, undefined, true)"
+  (legendLabelDeactivate)="onDeactivate($event, undefined, true)"
+  (legendLabelClick)="onClick($event)">
+  <svg:g [attr.transform]="transform" class="bar-chart chart">
+    <svg:g
+      lcu-charts-x-axis
+      *ngIf="xAxis"
+      [xScale]="xScale"
+      [dims]="dims"
+      [showLabel]="showXAxisLabel"
+      [labelText]="xAxisLabel"
+      [trimTicks]="trimXAxisTicks"
+      [rotateTicks]="rotateXAxisTicks"
+      [maxTickLength]="maxXAxisTickLength"
+      [tickFormatting]="xAxisTickFormatting"
+      [ticks]="xAxisTicks"
+      [xAxisOffset]="dataLabelMaxHeight.negative"
+      (dimensionsChanged)="updateXAxisHeight($event)"
+    ></svg:g>
+    <svg:g
+      lcu-charts-y-axis
+      *ngIf="yAxis"
+      [yScale]="yScale"
+      [dims]="dims"
+      [showGridLines]="showGridLines"
+      [showLabel]="showYAxisLabel"
+      [labelText]="yAxisLabel"
+      [trimTicks]="trimYAxisTicks"
+      [maxTickLength]="maxYAxisTickLength"
+      [tickFormatting]="yAxisTickFormatting"
+      [ticks]="yAxisTicks"
+      (dimensionsChanged)="updateYAxisWidth($event)"
+    ></svg:g>
+    <svg:g
+      *ngFor="let group of results; let index = index; trackBy: trackBy"
+      [@animationState]="'active'"
+      [attr.transform]="groupTransform(group)"
+    >
+      <svg:g
+        lcu-charts-series-vertical
+        type="stacked"
+        [xScale]="xScale"
+        [yScale]="yScale"
+        [activeEntries]="activeEntries"
+        [colors]="colors"
+        [series]="group.series"
+        [dims]="dims"
+        [gradient]="gradient"
+        [tooltipDisabled]="tooltipDisabled"
+        [tooltipTemplate]="tooltipTemplate"
+        [showDataLabel]="showDataLabel"
+        [dataLabelFormatting]="dataLabelFormatting"
+        [seriesName]="group.name"
+        [animations]="animations"
+        [noBarWhenZero]="noBarWhenZero"
+        (select)="onClick($event, group)"
+        (activate)="onActivate($event, group)"
+        (deactivate)="onDeactivate($event, group)"
+        (dataLabelHeightChanged)="onDataLabelMaxHeightChanged($event, index)"
+      />
+    </svg:g>
+  </svg:g>
+</lcu-charts-chart>
