+++ projects/common/src/lib/controls/bar-chart/bar-vertical-stacked/bar-vertical-stacked.component.ts
@@ -2,93 +2,25 @@ import {
   Component,
   Input,
   Output,
-  ViewEncapsulation,
   EventEmitter,
+  ViewEncapsulation,
   ChangeDetectionStrategy,
   ContentChild,
   TemplateRef
 } from '@angular/core';
 import { trigger, style, animate, transition } from '@angular/animations';
-
 import { scaleBand, scaleLinear } from 'd3-scale';
 
-import { calculateViewDimensions, ViewDimensions } from '../../common/view-dimensions.helper';
-import { ColorHelper } from '../../common/color.helper';
-import { BaseChartComponent } from '../../common/base-chart.component';
+import { calculateViewDimensions, ViewDimensions } from '../../../common/view-dimensions.helper';
+import { ColorHelper } from '../../../common/color.helper';
+import { BaseChartComponent } from '../../../common/base-chart.component';
 
 @Component({
-  selector: 'lcu-charts-bar-horizontal-normalized',
-  template: `
-    <lcu-charts-chart
-      [view]="[width, height]"
-      [showLegend]="legend"
-      [legendOptions]="legendOptions"
-      [activeEntries]="activeEntries"
-      [animations]="animations"
-      (legendLabelActivate)="onActivate($event, undefined, true)"
-      (legendLabelDeactivate)="onDeactivate($event, undefined, true)"
-      (legendLabelClick)="onClick($event)"
-    >
-      <svg:g [attr.transform]="transform" class="bar-chart chart">
-        <svg:g
-          lcu-charts-x-axis
-          *ngIf="xAxis"
-          [xScale]="xScale"
-          [dims]="dims"
-          [showGridLines]="showGridLines"
-          [showLabel]="showXAxisLabel"
-          [labelText]="xAxisLabel"
-          [trimTicks]="trimXAxisTicks"
-          [rotateTicks]="rotateXAxisTicks"
-          [maxTickLength]="maxXAxisTickLength"
-          [tickFormatting]="xAxisTickFormatting"
-          [ticks]="xAxisTicks"
-          (dimensionsChanged)="updateXAxisHeight($event)"
-        ></svg:g>
-        <svg:g
-          lcu-charts-y-axis
-          *ngIf="yAxis"
-          [yScale]="yScale"
-          [dims]="dims"
-          [showLabel]="showYAxisLabel"
-          [labelText]="yAxisLabel"
-          [trimTicks]="trimYAxisTicks"
-          [maxTickLength]="maxYAxisTickLength"
-          [tickFormatting]="yAxisTickFormatting"
-          [ticks]="yAxisTicks"
-          (dimensionsChanged)="updateYAxisWidth($event)"
-        ></svg:g>
-        <svg:g
-          *ngFor="let group of results; trackBy: trackBy"
-          [@animationState]="'active'"
-          [attr.transform]="groupTransform(group)"
-        >
-          <svg:g
-            lcu-charts-series-horizontal
-            type="normalized"
-            [xScale]="xScale"
-            [yScale]="yScale"
-            [activeEntries]="activeEntries"
-            [colors]="colors"
-            [series]="group.series"
-            [dims]="dims"
-            [gradient]="gradient"
-            [tooltipDisabled]="tooltipDisabled"
-            [tooltipTemplate]="tooltipTemplate"
-            [seriesName]="group.name"
-            [animations]="animations"
-            (select)="onClick($event, group)"
-            (activate)="onActivate($event, group)"
-            (deactivate)="onDeactivate($event, group)"
-            [noBarWhenZero]="noBarWhenZero"
-          />
-        </svg:g>
-      </svg:g>
-    </lcu-charts-chart>
-  `,
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  styleUrls: ['../../common/base-chart.component.scss'],
+  selector: 'lcu-charts-bar-vertical-stacked',
+  templateUrl: './bar-vertical-stacked.component.html',
+  styleUrls: ['../../../common/base-chart.component.scss'],
   encapsulation: ViewEncapsulation.None,
+  changeDetection: ChangeDetectionStrategy.OnPush,
   animations: [
     trigger('animationState', [
       transition(':leave', [
@@ -101,7 +33,7 @@ import { BaseChartComponent } from '../../common/base-chart.component';
     ])
   ]
 })
-export class BarHorizontalNormalizedComponent extends BaseChartComponent {
+export class BarVerticalStackedComponent extends BaseChartComponent {
   @Input() legend = false;
   @Input() legendTitle: string = 'Legend';
   @Input() legendPosition: string = 'right';
@@ -127,6 +59,9 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
   @Input() yAxisTicks: any[];
   @Input() barPadding = 8;
   @Input() roundDomains: boolean = false;
+  @Input() yScaleMax: number;
+  @Input() showDataLabel: boolean = false;
+  @Input() dataLabelFormatting: any;
   @Input() noBarWhenZero: boolean = true;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
@@ -141,14 +76,21 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
   xScale: any;
   yScale: any;
   transform: string;
+  tickFormatting: (label: string) => string;
   colors: ColorHelper;
   margin = [10, 20, 10, 20];
   xAxisHeight: number = 0;
   yAxisWidth: number = 0;
   legendOptions: any;
+  dataLabelMaxHeight: any = { negative: 0, positive: 0 };
 
   update(): void {
     super.update();
+
+    if (!this.showDataLabel) {
+      this.dataLabelMaxHeight = { negative: 0, positive: 0 };
+    }
+    this.margin = [10 + this.dataLabelMaxHeight.positive, 20, 10 + this.dataLabelMaxHeight.negative, 20];
 
     this.dims = calculateViewDimensions({
       width: this.width,
@@ -165,6 +107,10 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
       legendPosition: this.legendPosition
     });
 
+    if (this.showDataLabel) {
+      this.dims.height -= this.dataLabelMaxHeight.negative;
+    }
+
     this.formatDates();
 
     this.groupDomain = this.getGroupDomain();
@@ -177,24 +123,21 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
     this.setColors();
     this.legendOptions = this.getLegendOptions();
 
-    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0]})`;
+    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0] + this.dataLabelMaxHeight.negative})`;
   }
 
-  getGroupDomain(): any[] {
+  getGroupDomain() {
     const domain = [];
-
     for (const group of this.results) {
       if (!domain.includes(group.label)) {
         domain.push(group.label);
       }
     }
-
     return domain;
   }
 
-  getInnerDomain(): any[] {
+  getInnerDomain() {
     const domain = [];
-
     for (const group of this.results) {
       for (const d of group.series) {
         if (!domain.includes(d.label)) {
@@ -202,35 +145,67 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
         }
       }
     }
-
     return domain;
   }
 
-  getValueDomain(): any[] {
-    return [0, 100];
+  getValueDomain() {
+    const domain = [];
+    let smallest = 0;
+    let biggest = 0;
+    for (const group of this.results) {
+      let smallestSum = 0;
+      let biggestSum = 0;
+      for (const d of group.series) {
+        if (d.value < 0) {
+          smallestSum += d.value;
+        } else {
+          biggestSum += d.value;
+        }
+        smallest = d.value < smallest ? d.value : smallest;
+        biggest = d.value > biggest ? d.value : biggest;
+      }
+      domain.push(smallestSum);
+      domain.push(biggestSum);
+    }
+    domain.push(smallest);
+    domain.push(biggest);
+
+    const min = Math.min(0, ...domain);
+    const max = this.yScaleMax ? Math.max(this.yScaleMax, ...domain) : Math.max(...domain);
+    return [min, max];
   }
 
-  getYScale(): any {
-    const spacing = this.groupDomain.length / (this.dims.height / this.barPadding + 1);
-
+  getXScale(): any {
+    const spacing = this.groupDomain.length / (this.dims.width / this.barPadding + 1);
     return scaleBand()
-      .rangeRound([0, this.dims.height])
+      .rangeRound([0, this.dims.width])
       .paddingInner(spacing)
       .domain(this.groupDomain);
   }
 
-  getXScale(): any {
+  getYScale(): any {
     const scale = scaleLinear()
-      .range([0, this.dims.width])
+      .range([this.dims.height, 0])
       .domain(this.valueDomain);
     return this.roundDomains ? scale.nice() : scale;
   }
 
-  groupTransform(group): string {
-    return `translate(0, ${this.yScale(group.name)})`;
+  onDataLabelMaxHeightChanged(event, groupIndex) {
+    if (event.size.negative) {
+      this.dataLabelMaxHeight.negative = Math.max(this.dataLabelMaxHeight.negative, event.size.height);
+    } else {
+      this.dataLabelMaxHeight.positive = Math.max(this.dataLabelMaxHeight.positive, event.size.height);
+    }
+    if (groupIndex === this.results.length - 1) {
+      setTimeout(() => this.update());
+    }
   }
 
-  onClick(data, group?): void {
+  groupTransform(group) {
+    return `translate(${this.xScale(group.name) || 0}, 0)`;
+  }
+
+  onClick(data, group?) {
     if (group) {
       data.series = group.name;
     }
@@ -238,7 +213,7 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
     this.select.emit(data);
   }
 
-  trackBy(index, item): string {
+  trackBy(index, item) {
     return item.name;
   }
 
@@ -273,12 +248,12 @@ export class BarHorizontalNormalizedComponent extends BaseChartComponent {
     return opts;
   }
 
-  updateYAxisWidth({ width }): void {
+  updateYAxisWidth({ width }) {
     this.yAxisWidth = width;
     this.update();
   }
 
-  updateXAxisHeight({ height }): void {
+  updateXAxisHeight({ height }) {
     this.xAxisHeight = height;
     this.update();
   }
