+++ projects/common/src/lib/controls/line-chart/line-simple/line-chart-simple.component.ts
@@ -9,171 +9,55 @@ import {
   ContentChild,
   TemplateRef
 } from '@angular/core';
-import { scaleLinear, scalePoint, scaleTime } from 'd3-scale';
+import { trigger, style, animate, transition } from '@angular/animations';
+import { scaleLinear, scaleTime, scalePoint } from 'd3-scale';
 import { curveLinear } from 'd3-shape';
-
-import { calculateViewDimensions, ViewDimensions } from '../../common/view-dimensions.helper';
-import { ColorHelper } from '../../common/color.helper';
-import { BaseChartComponent } from '../../common/base-chart.component';
-import { id } from '../../utils/id';
-import { getUniqueXDomainValues, getScaleType } from '../../common/domain.helper';
+import { calculateViewDimensions, ViewDimensions } from '../../../common/view-dimensions.helper';
+import { ColorHelper } from '../../../common/color.helper';
+import { BaseChartComponent } from '../../../common/base-chart.component';
+import { id } from '../../../utils/id';
+import { getUniqueXDomainValues, getScaleType } from '../../../common/domain.helper';
 
 @Component({
-  selector: 'lcu-charts-area-chart',
-  template: `
-    <lcu-charts-chart
-      [view]="[width, height]"
-      [showLegend]="legend"
-      [legendOptions]="legendOptions"
-      [activeEntries]="activeEntries"
-      [animations]="animations"
-      (legendLabelClick)="onClick($event)"
-      (legendLabelActivate)="onActivate($event)"
-      (legendLabelDeactivate)="onDeactivate($event)"
-    >
-      <svg:defs>
-        <svg:clipPath [attr.id]="clipPathId">
-          <svg:rect
-            [attr.width]="dims.width + 10"
-            [attr.height]="dims.height + 10"
-            [attr.transform]="'translate(-5, -5)'"
-          />
-        </svg:clipPath>
-      </svg:defs>
-      <svg:g [attr.transform]="transform" class="area-chart chart">
-        <svg:g
-          lcu-charts-x-axis
-          *ngIf="xAxis"
-          [xScale]="xScale"
-          [dims]="dims"
-          [showGridLines]="showGridLines"
-          [showLabel]="showXAxisLabel"
-          [labelText]="xAxisLabel"
-          [trimTicks]="trimXAxisTicks"
-          [rotateTicks]="rotateXAxisTicks"
-          [maxTickLength]="maxXAxisTickLength"
-          [tickFormatting]="xAxisTickFormatting"
-          [ticks]="xAxisTicks"
-          (dimensionsChanged)="updateXAxisHeight($event)"
-        ></svg:g>
-        <svg:g
-          lcu-charts-y-axis
-          *ngIf="yAxis"
-          [yScale]="yScale"
-          [dims]="dims"
-          [showGridLines]="showGridLines"
-          [showLabel]="showYAxisLabel"
-          [labelText]="yAxisLabel"
-          [trimTicks]="trimYAxisTicks"
-          [maxTickLength]="maxYAxisTickLength"
-          [tickFormatting]="yAxisTickFormatting"
-          [ticks]="yAxisTicks"
-          (dimensionsChanged)="updateYAxisWidth($event)"
-        ></svg:g>
-        <svg:g [attr.clip-path]="clipPath">
-          <svg:g *ngFor="let series of results; trackBy: trackBy">
-            <svg:g
-              lcu-charts-area-series
-              [xScale]="xScale"
-              [yScale]="yScale"
-              [baseValue]="baseValue"
-              [colors]="colors"
-              [data]="series"
-              [activeEntries]="activeEntries"
-              [scaleType]="scaleType"
-              [gradient]="gradient"
-              [curve]="curve"
-              [animations]="animations"
-            />
-          </svg:g>
-
-          <svg:g *ngIf="!tooltipDisabled" (mouseleave)="hideCircles()">
-            <svg:g
-              lcu-charts-tooltip-area
-              [dims]="dims"
-              [xSet]="xSet"
-              [xScale]="xScale"
-              [yScale]="yScale"
-              [results]="results"
-              [colors]="colors"
-              [tooltipDisabled]="tooltipDisabled"
-              [tooltipTemplate]="seriesTooltipTemplate"
-              (hover)="updateHoveredVertical($event)"
-            />
-
-            <svg:g *ngFor="let series of results">
-              <svg:g
-                lcu-charts-circle-series
-                [xScale]="xScale"
-                [yScale]="yScale"
-                [colors]="colors"
-                [activeEntries]="activeEntries"
-                [data]="series"
-                [scaleType]="scaleType"
-                [visibleValue]="hoveredVertical"
-                [tooltipDisabled]="tooltipDisabled"
-                [tooltipTemplate]="tooltipTemplate"
-                (select)="onClick($event, series)"
-                (activate)="onActivate($event)"
-                (deactivate)="onDeactivate($event)"
-              />
-            </svg:g>
-          </svg:g>
-        </svg:g>
-      </svg:g>
-      <svg:g
-        lcu-charts-timeline
-        *ngIf="timeline && scaleType != 'ordinal'"
-        [attr.transform]="timelineTransform"
-        [results]="results"
-        [view]="[timelineWidth, height]"
-        [height]="timelineHeight"
-        [scheme]="scheme"
-        [customColors]="customColors"
-        [legend]="legend"
-        [scaleType]="scaleType"
-        (onDomainChange)="updateDomain($event)"
-      >
-        <svg:g *ngFor="let series of results; trackBy: trackBy">
-          <svg:g
-            lcu-charts-area-series
-            [xScale]="timelineXScale"
-            [yScale]="timelineYScale"
-            [baseValue]="baseValue"
-            [colors]="colors"
-            [data]="series"
-            [scaleType]="scaleType"
-            [gradient]="gradient"
-            [curve]="curve"
-            [animations]="animations"
-          />
-        </svg:g>
-      </svg:g>
-    </lcu-charts-chart>
-  `,
+  selector: 'lcu-charts-line-chart-simple',
+  templateUrl: './line-chart-simple.component.html',
+  styleUrls: ['../../../common/base-chart.component.scss'],
+  encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  styleUrls: ['../../common/base-chart.component.scss'],
-  encapsulation: ViewEncapsulation.None
+  animations: [
+    trigger('animationState', [
+      transition(':leave', [
+        style({
+          opacity: 1
+        }),
+        animate(
+          500,
+          style({
+            opacity: 0
+          })
+        )
+      ])
+    ])
+  ]
 })
-export class AreaChartComponent extends BaseChartComponent {
+export class LineChartSimpleComponent extends BaseChartComponent {
   @Input() legend;
   @Input() legendTitle: string = 'Legend';
   @Input() legendPosition: string = 'right';
-  @Input() state;
   @Input() xAxis;
   @Input() yAxis;
-  @Input() baseValue: any = 'auto';
-  @Input() autoScale;
   @Input() showXAxisLabel;
   @Input() showYAxisLabel;
   @Input() xAxisLabel;
   @Input() yAxisLabel;
+  @Input() autoScale;
   @Input() timeline;
   @Input() gradient: boolean;
   @Input() showGridLines: boolean = true;
   @Input() curve: any = curveLinear;
   @Input() activeEntries: any[] = [];
   @Input() schemeType: string;
+  @Input() rangeFillOpacity: number;
   @Input() trimXAxisTicks: boolean = true;
   @Input() trimYAxisTicks: boolean = true;
   @Input() rotateXAxisTicks: boolean = true;
@@ -185,37 +69,53 @@ export class AreaChartComponent extends BaseChartComponent {
   @Input() yAxisTicks: any[];
   @Input() roundDomains: boolean = false;
   @Input() tooltipDisabled: boolean = false;
+  @Input() showRefLines: boolean = false;
+  @Input() referenceLines: any;
+  @Input() showRefLabels: boolean = true;
   @Input() xScaleMin: any;
   @Input() xScaleMax: any;
   @Input() yScaleMin: number;
   @Input() yScaleMax: number;
+  @Input() backgroundGradientConfigs: any[];
+
+  @Input() xAxisIsDate: boolean;
+  /**
+   * example {DayOfWeek: true,
+   *          Month: false,
+   *          DayOfMonth: true,
+   *          Year: false,
+   *          Time: true,
+   *          TimeZone: false}
+   */
+  @Input() xAxisDateFormat: object;
 
   @Output() activate: EventEmitter<any> = new EventEmitter();
   @Output() deactivate: EventEmitter<any> = new EventEmitter();
 
-  @ContentChild('tooltipTemplate', { static: false }) tooltipTemplate: TemplateRef<any>;
-  @ContentChild('seriesTooltipTemplate', { static: false }) seriesTooltipTemplate: TemplateRef<any>;
+  @ContentChild('tooltipTemplate', {static: false}) tooltipTemplate: TemplateRef<any>;
+  @ContentChild('seriesTooltipTemplate', {static: false}) seriesTooltipTemplate: TemplateRef<any>;
 
   dims: ViewDimensions;
   xSet: any;
   xDomain: any;
   yDomain: any;
   seriesDomain: any;
-  xScale: any;
   yScale: any;
-  transform: string;
+  xScale: any;
   colors: ColorHelper;
-  clipPathId: string;
-  clipPath: string;
   scaleType: string;
+  transform: string;
+  clipPath: string;
+  clipPathId: string;
   series: any;
+  areaPath: any;
   margin = [10, 20, 10, 20];
   hoveredVertical: any; // the value of the x axis that is hovered over
   xAxisHeight: number = 0;
   yAxisWidth: number = 0;
   filteredDomain: any;
   legendOptions: any;
-
+  hasRange: boolean; // whether the line has a min-max range around it
   timelineWidth: any;
   timelineHeight: number = 50;
   timelineXScale: any;
@@ -241,6 +141,12 @@ export class AreaChartComponent extends BaseChartComponent {
       legendType: this.schemeType,
       legendPosition: this.legendPosition
     });
+/**\
+ * if the X axis is a date and they passed in a format
+ */
+    // if(this.xAxisIsDate && this.xAxisDateFormat){
+    //   this.xAxisTickFormatting = this.formatXAxisDate.bind(this);
+    // }
 
     if (this.timeline) {
       this.dims.height -= this.timelineHeight + this.margin[2] + this.timelinePadding;
@@ -262,10 +168,19 @@ export class AreaChartComponent extends BaseChartComponent {
     this.setColors();
     this.legendOptions = this.getLegendOptions();
 
-    this.transform = `translate(${this.dims.xOffset}, ${this.margin[0]})`;
+    this.transform = `translate(${this.dims.xOffset} , ${this.margin[0]})`;
 
     this.clipPathId = 'clip' + id().toString();
     this.clipPath = `url(#${this.clipPathId})`;
+  }
+
+  formatXAxisDate(value: Date){
+    console.log("value... = ", value);
+    // let newDate;
+    // if(this.xAxisDateFormat.DayOfWeek)
+    // this.xAxisDateFormat
+  let dateTime = value.toString();
+    return dateTime;
   }
 
   updateTimeline(): void {
@@ -319,11 +234,22 @@ export class AreaChartComponent extends BaseChartComponent {
 
   getYDomain(): any[] {
     const domain = [];
-
     for (const results of this.results) {
       for (const d of results.series) {
-        if (!domain.includes(d.value)) {
+        if (domain.indexOf(d.value) < 0) {
           domain.push(d.value);
+        }
+        if (d.min !== undefined) {
+          this.hasRange = true;
+          if (domain.indexOf(d.min) < 0) {
+            domain.push(d.min);
+          }
+        }
+        if (d.max !== undefined) {
+          this.hasRange = true;
+          if (domain.indexOf(d.max) < 0) {
+            domain.push(d.max);
+          }
         }
       }
     }
@@ -331,9 +257,6 @@ export class AreaChartComponent extends BaseChartComponent {
     const values = [...domain];
     if (!this.autoScale) {
       values.push(0);
-    }
-    if (this.baseValue !== 'auto') {
-      values.push(this.baseValue);
     }
 
     const min = this.yScaleMin ? this.yScaleMin : Math.min(...values);
@@ -351,54 +274,33 @@ export class AreaChartComponent extends BaseChartComponent {
     let scale;
 
     if (this.scaleType === 'time') {
-      scale = scaleTime();
+      scale = scaleTime()
+        .range([0, width])
+        .domain(domain);
     } else if (this.scaleType === 'linear') {
-      scale = scaleLinear();
+      scale = scaleLinear()
+        .range([0, width])
+        .domain(domain);
+
+      if (this.roundDomains) {
+        scale = scale.nice();
+      }
     } else if (this.scaleType === 'ordinal') {
-      scale = scalePoint().padding(0.1);
+      scale = scalePoint()
+        .range([0, width])
+        .padding(0.1)
+        .domain(domain);
     }
 
-    scale.range([0, width]).domain(domain);
-
-    return this.roundDomains ? scale.nice() : scale;
+    return scale;
   }
 
   getYScale(domain, height): any {
     const scale = scaleLinear()
       .range([height, 0])
       .domain(domain);
+
     return this.roundDomains ? scale.nice() : scale;
-  }
-
-  getScaleType(values): string {
-    let date = true;
-    let num = true;
-    for (const value of values) {
-      if (!this.isDate(value)) {
-        date = false;
-      }
-      if (typeof value !== 'number') {
-        num = false;
-      }
-    }
-
-    if (date) {
-      return 'time';
-    }
-
-    if (num) {
-      return 'linear';
-    }
-
-    return 'ordinal';
-  }
-
-  isDate(value): boolean {
-    if (value instanceof Date) {
-      return true;
-    }
-
-    return false;
   }
 
   updateDomain(domain): void {
@@ -418,11 +320,7 @@ export class AreaChartComponent extends BaseChartComponent {
     this.deactivateAll();
   }
 
-  onClick(data, series?): void {
-    if (series) {
-      data.series = series.name;
-    }
-
+  onClick(data): void {
     this.select.emit(data);
   }
 
@@ -471,6 +369,8 @@ export class AreaChartComponent extends BaseChartComponent {
   }
 
   onActivate(item) {
+    this.deactivateAll();
+
     const idx = this.activeEntries.findIndex(d => {
       return d.name === item.name && d.value === item.value;
     });
@@ -478,7 +378,7 @@ export class AreaChartComponent extends BaseChartComponent {
       return;
     }
 
-    this.activeEntries = [item, ...this.activeEntries];
+    this.activeEntries = [item];
     this.activate.emit({ value: item, entries: this.activeEntries });
   }
