+++ projects/demo/src/app/charts/demo-area-chart/demo-area-chart.component.html
@@ -1,4 +1,4 @@
-<lcu-charts-area-chart
+<lcu-charts-area-chart-simple
   class="chart-container"
   [animations]="animations"
   [autoScale]="autoScale"
@@ -34,4 +34,4 @@
   (deactivate)="deactivate($event)"
   (legendLabelClick)="onLegendLabelClick($event)"
   (select)="select($event)">
-</lcu-charts-area-chart>
+</lcu-charts-area-chart-simple>
