+++ projects/demo/src/app/charts/demo-line-chart/demo-line-chart.component.html
@@ -1,8 +1,8 @@
-<lcu-charts-line-chart
+<lcu-charts-line-chart-simple
   class="chart-container"
-  [backgroundGradientConfigs]="backgroundGradientConfigs"
   [animations]="animations"
   [autoScale]="autoScale"
+  [backgroundGradientConfigs]="backgroundGradientConfigs"
   [curve]="curve"
   [gradient]="gradient"
   [legendPosition]="legendPosition"
@@ -24,18 +24,16 @@
   [trimXAxisTicks]="trimXAxisTicks"
   [trimYAxisTicks]="trimYAxisTicks"
   [view]="view"
+  [xAxisDateFormat]="xAxisDateFormat"
+  [xAxisIsDate]="xAxisIsDate"
   [xAxisLabel]="xAxisLabel"
+  [xAxisTickFormatting]="xAxisTickFormatting"
   [xAxis]="showXAxis"
   [xScaleMax]="xScaleMax"
   [xScaleMin]="xScaleMin"
   [yAxisLabel]="yAxisLabel"
-
   [yAxisTickFormatting]="yAxisTickFormatting"
-  [xAxisTickFormatting]="xAxisTickFormatting"
   [yAxisTicks]="yAxisTicks"
-  [xAxisIsDate]="xAxisIsDate"
-  [xAxisDateFormat]="xAxisDateFormat"
-
   [yAxis]="showYAxis"
   [yScaleMax]="yScaleMax"
   [yScaleMin]="yScaleMin"
@@ -43,4 +41,4 @@
   (deactivate)="deactivate($event)"
   (legendLabelClick)="onLegendLabelClick($event)"
   (select)="select($event)">
-</lcu-charts-line-chart>
+</lcu-charts-line-chart-simple>
