+++ projects/demo/src/app/charts/demo-vertical-bar-chart/demo-vertical-bar-chart.component.html
@@ -1,8 +1,7 @@
-<lcu-charts-bar-vertical
+<lcu-charts-bar-vertical-simple
   class="chart-container"
   [animations]="animations"
   [barPadding]="barPadding"
-  [customColors]="customColors"
   [gradient]="gradient"
   [legendPosition]="legendPosition"
   [legendTitle]="legendTitle"
@@ -33,4 +32,4 @@
   (deactivate)="deactivate($event)"
   (legendLabelClick)="onLegendLabelClick($event)"
   (select)="select($event)">
-</lcu-charts-bar-vertical>
+</lcu-charts-bar-vertical-simple>
